+++ CHANGELOG.md
@@ -5,7 +5,8 @@ ChangeLog
 0.9.6 (2026-06-07)
 ------------------
 
-- FEATURE: add `project.artifact.{spec,arch,soft,docs,infr}` config vars
+- FEATURE: add `ase_artifact_name(kind, name)` MCP tool and `ase artifact name --kind <kind> <name>` CLI to create artifact filenames
+- FEATURE: add `project.artifact.{spec,arch,soft,docs,infr}.{basedir,files}` config vars
 - FEATURE: add `ase_artifact_list(kind)` MCP tool and `ase artifact list --kind` CLI to resolve artifact kinds to project files
 - FEATURE: add missing `ase config delete` CLI sub-command to delete a configuration key